Fix build warnings on Solaris2
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@24814 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/rcfpch/rcf_pch_conf.c
@@ -1187,7 +1187,7 @@ create_lock(const char *name)
     {
         rc = TE_OS_RC(TE_RCF_PCH, errno);
     }
-    else if (fprintf(f, "%d", getpid()) < 0)
+    else if (fprintf(f, "%ld", (long)getpid()) < 0)
     {
         rc = TE_OS_RC(TE_RCF_PCH, errno);
         fclose(f);
